perf: use repository for me query & caching in /settings/my-account/general/ RSC
