+++ apps/web/app/(use-page-wrapper)/settings/(settings-layout)/my-account/general/actions.ts
@@ -1,0 +1,16 @@
+"use server";
+
+import { revalidateTag } from "next/cache";
+
+const revalidateGetMe = () => {
+  revalidateTag("viewer.me.get");
+};
+
+const revalidateTravelSchedules = () => {
+  revalidateTag("viewer.travelSchedules.get");
+};
+
+export const revalidatePage = async () => {
+  revalidateTravelSchedules();
+  revalidateGetMe();
+};
