+++ packages/lib/server/repository/secondaryEmail.ts
@@ -1,0 +1,16 @@
+import prisma from "@calcom/prisma";
+
+export class SecondaryEmailRepository {
+  static async findAllSecondaryEmailsByUserId(userId: number) {
+    return await prisma.secondaryEmail.findMany({
+      where: {
+        userId,
+      },
+      select: {
+        id: true,
+        email: true,
+        emailVerified: true,
+      },
+    });
+  }
+}
