+++ packages/lib/server/repository/user.ts
@@ -1,4 +1,4 @@
-import type { z } from "zod";
+import { z } from "zod";
 
 import { whereClauseForOrgWithSlugOrRequestedSlug } from "@calcom/ee/organizations/lib/orgDomains";
 import logger from "@calcom/lib/logger";
@@ -8,15 +8,20 @@ import prisma from "@calcom/prisma";
 import { Prisma } from "@calcom/prisma/client";
 import type { User as UserType } from "@calcom/prisma/client";
 import type { CreationSource } from "@calcom/prisma/enums";
-import { MembershipRole } from "@calcom/prisma/enums";
+import { IdentityProvider, MembershipRole } from "@calcom/prisma/enums";
 import { credentialForCalendarServiceSelect } from "@calcom/prisma/selects/credential";
-import { userMetadata } from "@calcom/prisma/zod-utils";
+import { teamMetadataSchema, userMetadata } from "@calcom/prisma/zod-utils";
 import type { UpId, UserProfile } from "@calcom/types/UserProfile";
 
 import { DEFAULT_SCHEDULE, getAvailabilityFromSchedule } from "../../availability";
+import { WEBAPP_URL } from "../../constants";
 import { buildNonDelegationCredentials } from "../../delegationCredential/clientAndServer";
+import { getUserAvatarUrl } from "../../getAvatarUrl";
 import { withSelectedCalendars } from "../withSelectedCalendars";
+import { AccountRepository } from "./account";
+import { MembershipRepository } from "./membership";
 import { ProfileRepository } from "./profile";
+import { SecondaryEmailRepository } from "./secondaryEmail";
 import { getParsedTeam } from "./teamUtils";
 
 export type { UserWithLegacySelectedCalendars } from "../withSelectedCalendars";
@@ -26,6 +31,14 @@ export type UserAdminTeams = number[];
 const log = logger.getSubLogger({ prefix: ["[repository/user]"] });
 
 export const ORGANIZATION_ID_UNKNOWN = "ORGANIZATION_ID_UNKNOWN";
+
+const ZMeOpts = z
+  .object({
+    includePasswordAdded: z.boolean().optional(),
+  })
+  .optional();
+
+type TMeOpts = z.infer<typeof ZMeOpts>;
 
 const teamSelect = Prisma.validator<Prisma.TeamSelect>()({
   id: true,
@@ -962,5 +975,212 @@ export class UserRepository {
       where: { id },
       data: { whitelistWorkflows },
     });
+  }
+
+  static async getUserFromSessionUpIdAndUserId({
+    upId,
+    userId,
+    url,
+  }: {
+    upId: string;
+    userId: number;
+    url?: string;
+  }) {
+    const userFromDb = await UserRepository.findUnlockedUserForSession({ userId });
+
+    if (!userFromDb) {
+      return null;
+    }
+
+    const user = await UserRepository.enrichUserWithTheProfile({
+      user: userFromDb,
+      upId,
+    });
+
+    if (url) {
+      logger.debug(
+        `getUserFromSession: enriched user with profile - ${url}`,
+        safeStringify({ user, userFromDb, upId })
+      );
+    } else {
+      logger.debug(
+        `getUserFromSessionUpIdAndUserId: enriched user with profile - ${safeStringify({
+          user,
+          userFromDb,
+          upId,
+        })}`
+      );
+    }
+
+    const { email, username, id } = user;
+    if (!email || !id) {
+      return null;
+    }
+
+    const userMetaData = userMetadata.parse(user.metadata || {});
+    const orgMetadata = teamMetadataSchema.parse(user.profile?.organization?.metadata || {});
+    // This helps to prevent reaching the 4MB payload limit by avoiding base64 and instead passing the avatar url
+
+    const locale = user?.locale;
+    const { members = [], ..._organization } = user.profile?.organization || {};
+    const isOrgAdmin = members.some((member) => ["OWNER", "ADMIN"].includes(member.role));
+
+    if (isOrgAdmin) {
+      logger.debug("User is an org admin", safeStringify({ userId: user.id }));
+    } else {
+      logger.debug("User is not an org admin", safeStringify({ userId: user.id }));
+    }
+    const organization = {
+      ..._organization,
+      id: user.profile?.organization?.id ?? null,
+      isOrgAdmin,
+      metadata: orgMetadata,
+      requestedSlug: orgMetadata?.requestedSlug ?? null,
+    };
+
+    return {
+      ...user,
+      avatar: `${WEBAPP_URL}/${user.username}/avatar.png?${organization.id}` && `orgId=${organization.id}`,
+      // TODO: OrgNewSchema - later -  We could consolidate the props in user.profile?.organization as organization is a profile thing now.
+      organization,
+      organizationId: organization.id,
+      id,
+      email,
+      username,
+      locale,
+      defaultBookerLayouts: userMetaData?.defaultBookerLayouts || null,
+    };
+  }
+
+  /**
+   * Retrieves the user profile and other related information for the current user.
+   * @param userId - The Id of user to retrieve, not necessary if passing sessionUser
+   * @param upId - The UpId of the user to retrieve
+   * @param sessionUser - If a user session is passed, then it is not fetched again and re used
+   */
+  static async getMe({
+    userId,
+    upId,
+    opts,
+    sessionUser,
+  }: {
+    userId: number;
+    upId: string;
+    opts?: TMeOpts;
+    sessionUser?: Awaited<ReturnType<typeof UserRepository.getUserFromSessionUpIdAndUserId>>;
+  }) {
+    const userFromSession =
+      sessionUser ||
+      (await UserRepository.getUserFromSessionUpIdAndUserId({
+        upId,
+        userId,
+      }));
+
+    if (!userFromSession) {
+      return null;
+    }
+
+    const allUserEnrichedProfiles = await ProfileRepository.findAllProfilesForUserIncludingMovedUser(
+      userFromSession
+    );
+
+    const user = await UserRepository.enrichUserWithTheProfile({
+      user: userFromSession,
+      upId,
+    });
+
+    const secondaryEmails = await SecondaryEmailRepository.findAllSecondaryEmailsByUserId(user.id);
+
+    let passwordAdded = false;
+    if (user.identityProvider !== IdentityProvider.CAL && opts?.includePasswordAdded) {
+      const userWithPassword = await prisma.user.findUnique({
+        where: {
+          id: user.id,
+        },
+        select: {
+          password: true,
+        },
+      });
+      if (userWithPassword?.password?.hash) {
+        passwordAdded = true;
+      }
+    }
+
+    let identityProviderEmail = "";
+    if (user.identityProviderId) {
+      const account = await AccountRepository.findUniqueByProviderAndProviderAccountId({
+        provider: user.identityProvider.toLocaleLowerCase(),
+        providerAccountId: user.identityProviderId,
+      });
+      identityProviderEmail = account?.providerEmail || "";
+    }
+
+    const userMetadataPrased = userMetadata.parse(user.metadata);
+
+    // Destructuring here only makes it more illegible
+    // pick only the part we want to expose in the API
+
+    const profileData = user.organization?.isPlatform
+      ? {
+          organizationId: null,
+          organization: { id: -1, isPlatform: true, slug: "", isOrgAdmin: false },
+          username: user.username ?? null,
+          profile: ProfileRepository.buildPersonalProfileFromUser({ user }),
+          profiles: [],
+        }
+      : {
+          organizationId: user.profile?.organizationId ?? null,
+          organization: user.organization,
+          username: user.profile?.username ?? user.username ?? null,
+          profile: user.profile ?? null,
+          profiles: allUserEnrichedProfiles,
+          organizationSettings: user?.profile?.organization?.organizationSettings,
+        };
+
+    const isTeamAdminOrOwner =
+      (await MembershipRepository.findFirstAdminOrOwnerMembershipByUserId({ userId: user.id })) !== null;
+
+    const crypto = await import("crypto");
+
+    return {
+      id: user.id,
+      name: user.name,
+      email: user.email,
+      emailMd5: crypto.createHash("md5").update(user.email).digest("hex"),
+      emailVerified: user.emailVerified,
+      startTime: user.startTime,
+      endTime: user.endTime,
+      bufferTime: user.bufferTime,
+      locale: user.locale,
+      timeFormat: user.timeFormat,
+      timeZone: user.timeZone,
+      avatar: getUserAvatarUrl(user),
+      avatarUrl: user.avatarUrl,
+      createdDate: user.createdDate,
+      trialEndsAt: user.trialEndsAt,
+      defaultScheduleId: user.defaultScheduleId,
+      completedOnboarding: user.completedOnboarding,
+      twoFactorEnabled: user.twoFactorEnabled,
+      disableImpersonation: user.disableImpersonation,
+      identityProvider: user.identityProvider,
+      identityProviderEmail,
+      brandColor: user.brandColor,
+      darkBrandColor: user.darkBrandColor,
+      bio: user.bio,
+      weekStart: user.weekStart,
+      theme: user.theme,
+      appTheme: user.appTheme,
+      hideBranding: user.hideBranding,
+      metadata: user.metadata,
+      defaultBookerLayouts: user.defaultBookerLayouts,
+      allowDynamicBooking: user.allowDynamicBooking,
+      allowSEOIndexing: user.allowSEOIndexing,
+      receiveMonthlyDigestEmail: user.receiveMonthlyDigestEmail,
+      ...profileData,
+      secondaryEmails,
+      isPremium: userMetadataPrased?.isPremium,
+      ...(passwordAdded ? { passwordAdded } : {}),
+      isTeamAdminOrOwner,
+    };
   }
 }
