+++ packages/lib/unstable_cache/index.ts
@@ -1,0 +1,3 @@
+import { cache as unstable_cache } from "./unstable_cache";
+
+export { unstable_cache };
